爬取百度文库的doc格式/抓取百度文库所有内容/带GUI的爬取百度文库/requirements.txt
pillow==6.1.0
requests
selenium
lxml
bs4